requirement.txt
python-dotenv==1.1.1
youtube-transcript-api==1.1.1
langchain==0.3.26
langchain-community==0.3.27
langchain-google-genai==2.1.7
faiss-cpu==1.11.0
langdetect==1.0.9